==> notebooks/analysis/set1/02_v2_compute_ECFP_n_final.ipynb <==
from rdkit.Chem.AllChem import GetMorganFingerprintAsBitVect
from rdkit.Chem import MolFromSmiles, MolToSmiles
from tqdm.auto import tqdm
tqdm.pandas()
import pandas as pd
import os
import pickle
from multiprocessing import Pool
from rdkit import RDLogger
import numpy as np
import itertools
RDLogger.DisableLog('rdApp.error')

# load datasets pickle from ../data/standardized_data

chemdiv_df = pd.read_pickle('../../../data/standardized_data/chemdiv_df.pkl')
enamine_df = pd.read_pickle('../../../data/standardized_data/enamine_rna_df.pkl')
enamine_protein_df = pd.read_pickle('../../../data/standardized_data/enamine_protein_df.pkl')
life_chemicals_df = pd.read_pickle('../../../data/standardized_data/life_chemicals_df.pkl')
robin_df = pd.read_pickle('../../../data/standardized_data/robin_df.pkl')



all_df = pd.concat([chemdiv_df, enamine_df, enamine_protein_df, life_chemicals_df, robin_df], ignore_index=True)
rna_df = pd.concat([chemdiv_df, enamine_df, life_chemicals_df, robin_df], ignore_index=True)

all_df

# number of molecules in each dataset
# sum without protein
print(f'Total ALL: {len(all_df)}')
print(f'Total RNA: {len(rna_df)}')
print(f'Enamine Protein: {len(enamine_protein_df)}')
print(f'Chemdiv: {len(chemdiv_df)}')
print(f'Enamine: {len(enamine_df)}')
print(f'Life Chemicals: {len(life_chemicals_df)}')
print(f'Robin: {len(robin_df)}')


# create a df where will be three columns, before, after and difference. Before deduplication, after deduplication and difference. Store the first column for each dataset
df_deduplication = pd.DataFrame(columns=['before', 'after', 'difference'])
df_deduplication['before'] = [len(enamine_protein_df), len(chemdiv_df), len(enamine_df), len(life_chemicals_df), len(robin_df)]
df_deduplication.index = ['Enamine Protein', 'Chemdiv', 'Enamine', 'Life Chemicals', 'Robin']
df_deduplication[['after', 'difference']] = 0

df_deduplication

# add column 'source' to each df
chemdiv_df['source'] = 'chemdiv'
enamine_df['source'] = 'enamine'
enamine_protein_df['source'] = 'enamine_protein'
life_chemicals_df['source'] = 'life_chemicals'
robin_df['source'] = 'robin'


# create SMILES column for each dataset using rdKit
datasets = [chemdiv_df, enamine_df, enamine_protein_df, life_chemicals_df, robin_df]

for dataset in datasets:
    dataset['SMILES'] = dataset['mol'].apply(MolToSmiles, canonical=True)


del rna_df, all_df

rna_df = pd.concat([chemdiv_df, enamine_df, life_chemicals_df, robin_df], ignore_index=True)
all_df = pd.concat([chemdiv_df, enamine_df, enamine_protein_df, life_chemicals_df, robin_df], ignore_index=True)

all_df.describe()

# import pandas as pd
# from rdkit import Chem
# from chembl_structure_pipeline import standardizer

# def standardize_smiles(smiles):
#     """Standardize SMILES using ChEMBL Structure Pipeline."""
#     mol = Chem.MolFromSmiles(smiles)
#     if mol:
#         standardized_mol = standardizer.standardize_mol(mol)
#         standardized_smiles = Chem.MolToSmiles(standardized_mol, canonical=True)
#         return standardized_smiles
#     return None

# def correct_smiles(smiles):
#     """Keep the longer part of the SMILES string, discarding salts."""
#     parts = smiles.split('.')
#     if len(parts) > 1:
#         largest_part = max(parts, key=lambda x: len(Chem.MolFromSmiles(x).GetAtoms()))
#         return largest_part
#     return smiles

# # Assuming all_df is your DataFrame containing molecules data
# # Remove the 'mol' column
# all_df.drop(columns=['mol'], inplace=True)

# # Initialize a list to store molecules with disconnected components
# disconnected_mols = []

# # Process each row in all_df
# for idx, row in all_df.iterrows():
#     smiles = row['SMILES']
#     source = row['source']

#     # Standardize and correct SMILES
#     standardized_smiles = standardize_smiles(smiles)
#     if standardized_smiles is None:
#         print(f"Invalid SMILES string at index {idx}: {smiles}")
#         continue

#     corrected_smiles = correct_smiles(standardized_smiles)
#     mol_from_smiles = Chem.MolFromSmiles(corrected_smiles)

#     if mol_from_smiles is None or mol_from_smiles.GetNumAtoms() == 0:
#         print(f"Invalid SMILES string after correction at index {idx}: {smiles}")
#         continue

#     num_components = Chem.GetMolFrags(mol_from_smiles, asMols=False, sanitizeFrags=False)
#     if len(num_components) > 1:
#         print(f"Disconnected components in molecule at index {idx}: {smiles}, from source {source}")
#         disconnected_mols.append({'Index': idx, 'SMILES': corrected_smiles, 'Source': source})
#         all_df.at[idx, 'SMILES'] = corrected_smiles  # Update SMILES with corrected version

# # Recompute 'mol' column from corrected SMILES
# all_df['mol'] = all_df['SMILES'].apply(lambda x: Chem.MolFromSmiles(x))

# # Convert the list to a DataFrame and save to JSON
# disconnected_mols_df = pd.DataFrame(disconnected_mols)
# disconnected_mols_df.to_json('disconnected_mols_df.json')

# # Optionally, print or return the DataFrame of disconnected molecules
# disconnected_mols_df

all_df

# check the SMILES column in all_df if they do not have dot in the string
dissconnected_molecules = all_df[all_df['SMILES'].str.contains('\.')]

# save the dissconnected molecules to json
dissconnected_molecules.to_json('mols_with_salt.json')


dissconnected_molecules

# from chembl_structure_pipeline import standardizer
# from rdkit import Chem

# def get_parent_mol(smiles):
#     """Remove salt from SMILES and return the parent molecule's SMILES."""
#     mol = Chem.MolFromSmiles(smiles)
#     if mol:
#         parent_mol, _ = standardizer.get_parent_mol(mol)  # Unpack the returned tuple
#         if parent_mol:  # Ensure parent_mol is not None
#             parent_smiles = Chem.MolToSmiles(parent_mol, canonical=True)
#             return parent_smiles
#     return None

# # Assuming dissconnected_molecules is a DataFrame containing molecules
# removed_salts = []
# for idx, row in dissconnected_molecules.iterrows():
#     smiles = row['SMILES']
#     source = row['source']
#     parent_smiles = get_parent_mol(smiles)
#     if parent_smiles is None:
#         print(f"Invalid SMILES string at index {idx}: {smiles}")
#         continue
#     removed_salts.append({'Index': idx, 'SMILES': parent_smiles, 'Source': source})
#     # Assuming you want to update all_df too
#     all_df.at[idx, 'SMILES'] = parent_smiles  # Update SMILES with corrected version

# # Process the removed_salts list as needed
# removed_salts_df = pd.DataFrame(removed_salts)
# removed_salts_df.to_json('removed_salts_df.json')



salt_check = all_df[all_df['SMILES'].str.contains('\.')].copy()
# drop mol column
salt_check


# # from salt_check, remove the [O-][Cl+3]([O-])([O-])[O-] from the SMILES
# salt_check['SMILES'] = salt_check['SMILES'].str.replace(r'[O-][Cl+3]([O-])([O-])[O-].', '')
# salt_check['SMILES'] = salt_check['SMILES'].str.replace(r'.[O-][Cl+3]([O-])([O-])[O-]', '')
# salt_check['SMILES'] = salt_check['SMILES'].str.replace(r'.[O-][Cl+3]([O-])([O-])O', '')


for x in salt_check['SMILES']:
    print(x)

all_df['SMILES']  = all_df['SMILES'].str.replace(r'[O-][Cl+3]([O-])([O-])[O-].', '')
all_df['SMILES']  = all_df['SMILES'].str.replace(r'.[O-][Cl+3]([O-])([O-])[O-]', '')
all_df['SMILES']  = all_df['SMILES'].str.replace(r'.[O-][Cl+3]([O-])([O-])O', '')

all_df[all_df['SMILES'].str.contains('\.')]

all_df.head(5)

# # Initialize a list to store molecules with disconnected components
# disconnected_mols = []

# # Process each row in all_df
# for idx, row in all_df.iterrows():
#     smiles = row['SMILES']
#     source = row['source']

#     # Standardize and correct SMILES
#     standardized_smiles = standardize_smiles(smiles)
#     if standardized_smiles is None:
#         print(f"Invalid SMILES string at index {idx}: {smiles}")
#         continue

#     corrected_smiles = correct_smiles(standardized_smiles)
#     mol_from_smiles = Chem.MolFromSmiles(corrected_smiles)

#     if mol_from_smiles is None or mol_from_smiles.GetNumAtoms() == 0:
#         print(f"Invalid SMILES string after correction at index {idx}: {smiles}")
#         continue

#     num_components = Chem.GetMolFrags(mol_from_smiles, asMols=False, sanitizeFrags=False)
#     if len(num_components) > 1:
#         print(f"Disconnected components in molecule at index {idx}: {smiles}, from source {source}")
#         disconnected_mols.append({'Index': idx, 'SMILES': corrected_smiles, 'Source': source})
#         all_df.at[idx, 'SMILES'] = corrected_smiles  # Update SMILES with corrected version

# # Recompute 'mol' column from corrected SMILES
# all_df['mol'] = all_df['SMILES'].apply(lambda x: Chem.MolFromSmiles(x))

# # Convert the list to a DataFrame and save to JSON
# disconnected_mols_2_df = pd.DataFrame(disconnected_mols)
# disconnected_mols_2_df.to_json('disconnected_mols_2_df.json')

# # Optionally, print or return the DataFrame of disconnected molecules
# disconnected_mols_2_df

all_df[all_df['SMILES'].str.contains('\.')]

# using source column, divide the all_df into 5 different dataframes
chemdiv_df = all_df[all_df['source'] == 'chemdiv']
enamine_df = all_df[all_df['source'] == 'enamine']
enamine_protein_df = all_df[all_df['source'] == 'enamine_protein']
life_chemicals_df = all_df[all_df['source'] == 'life_chemicals']
robin_df = all_df[all_df['source'] == 'robin']

# statistics of the number of molecules in each dataset
print(f'Chemdiv: {len(chemdiv_df)}')
print(f'Enamine: {len(enamine_df)}')
print(f'Enamine Protein: {len(enamine_protein_df)}')
print(f'Life Chemicals: {len(life_chemicals_df)}')
print(f'Robin: {len(robin_df)}')


# List of datasets
datasets = {
    'Enamine protein': enamine_protein_df,
    'Chemdiv': chemdiv_df,
    'Enamine': enamine_df,
    'Life Chemicals': life_chemicals_df,
    'ROBIN': robin_df,
}

# remove duplicates within datasets
for dataset in datasets:
    datasets[dataset] = datasets[dataset].drop_duplicates(subset='SMILES', keep='first')
    

# create a df to store the number of duplicates between each pair of datasets
df_dup_count = pd.DataFrame(columns=datasets.keys(), index=datasets.keys())
df_dup_count = df_dup_count.fillna(0)
    
# Iterate over all pairs of datasets and count duplicates
for (name1, df1), (name2, df2) in itertools.combinations(datasets.items(), 2):
    # Find duplicates between df1 and df2
    duplicates = df1[df1['SMILES'].isin(df2['SMILES'])]
    
    # Store the number of duplicates in the df
    df_dup_count.loc[name1, name2] = len(duplicates)
    df_dup_count.loc[name2, name1] = len(duplicates)
    
    # Print the count of duplicates
    print(f"Duplicates between {name1} and {name2}: {len(duplicates)}")
    

# center the diagonal
df_dup_count.values[[np.arange(len(df_dup_count))]*2] = 0


    
# drop duplicates in rna_df and enamine_protein_df
rna_df = rna_df.drop_duplicates(subset='SMILES', keep='first')
enamine_protein_df = enamine_protein_df.drop_duplicates(subset='SMILES', keep='first')

rna_prot_dup = rna_df[rna_df['SMILES'].isin(enamine_protein_df['SMILES'])]
print(f'Duplicates between RNA and Enamine Protein: {len(rna_prot_dup)}')
df_dup_count


print(f'Chemdiv: {len(chemdiv_df)}')
print(f'Enamine: {len(enamine_df)}')
print(f'Enamine Protein: {len(enamine_protein_df)}')
print(f'Life Chemicals: {len(life_chemicals_df)}')
print(f'Robin: {len(robin_df)}')

# # Save the number of duplicates to a pickle file
# df_dup_count.to_pickle('duplicate_counts_v4.pkl')
    
# # Save the number of duplicates to a text file
# duplicate_counts = df_dup_count.to_dict()
# with open('duplicate_counts_v4.txt', 'w') as f:
#     for key, value in duplicate_counts.items():
#         f.write(f'{key}: {value}\n')
    
# # save found duplicates to a pickle file
# duplicates = rna_df[rna_df['SMILES'].isin(enamine_protein_df['SMILES'])]
# duplicates.to_pickle('duplicates_v4.pkl')


# # Combine datasets into one DataFrame
# combined_df = pd.concat([enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df])

# # Sort datasets by size (largest to smallest)
# datasets = [enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df]
# sorted_datasets = sorted(datasets, key=lambda x: x.shape[0], reverse=True)

# # Finding all unique duplicates across datasets
# all_duplicates = combined_df[combined_df.duplicated('SMILES', keep=False)]

# # Remove duplicates in order from largest to smallest, except the smallest
# for i in range(len(sorted_datasets) - 1):  # Skip the smallest dataset
#     dataset = sorted_datasets[i]
#     duplicates_to_remove = all_duplicates[~all_duplicates['source'].isin([dataset['source'].iloc[0]])]['SMILES']
#     sorted_datasets[i] = dataset[~dataset['SMILES'].isin(duplicates_to_remove)]

# # Extracting the updated datasets
# enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df = sorted_datasets

# # Now the datasets are deduplicated in the desired order

## remove all duplicates





# Combine datasets into one DataFrame
combined_df = pd.concat([enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df])

# Finding all duplicates across datasets
all_duplicates = combined_df[combined_df.duplicated('SMILES', keep=False)]['SMILES'].unique()

# Remove these duplicates from each dataset
enamine_protein_df = enamine_protein_df[~enamine_protein_df['SMILES'].isin(all_duplicates)]
chemdiv_df = chemdiv_df[~chemdiv_df['SMILES'].isin(all_duplicates)]
enamine_df = enamine_df[~enamine_df['SMILES'].isin(all_duplicates)]
life_chemicals_df = life_chemicals_df[~life_chemicals_df['SMILES'].isin(all_duplicates)]
robin_df = robin_df[~robin_df['SMILES'].isin(all_duplicates)]

# Check for and remove duplicates within each individual dataset
datasets = [enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df]
for i, dataset in enumerate(datasets):
    datasets[i] = dataset.drop_duplicates(subset='SMILES')

# Extracting the updated datasets
enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df = datasets

# # find duplicates in enamine_protein_df and remove them

# enamine_protein_df = enamine_protein_df.drop_duplicates(subset='SMILES', keep='first')


# List of datasets
datasets = {
    'enamine_protein': enamine_protein_df,
    'chemdiv': chemdiv_df,
    'enamine': enamine_df,
    'life_chemicals': life_chemicals_df,
    'robin': robin_df
}

# Iterate over all pairs of datasets and count duplicates
for (name1, df1), (name2, df2) in itertools.combinations(datasets.items(), 2):
    # Find duplicates between df1 and df2
    duplicates = df1[df1['SMILES'].isin(df2['SMILES'])]
    
    # Print the count of duplicates
    print(f"Duplicates between {name1} and {name2}: {len(duplicates)}")

df_deduplication['after'] = [len(enamine_protein_df), len(chemdiv_df), len(enamine_df), len(life_chemicals_df), len(robin_df)]
df_deduplication['difference'] = df_deduplication['before'] - df_deduplication['after']
df_deduplication

# Load your datasets (assuming they are already loaded as per your previous message)

# Verify and sort datasets by size
datasets = [enamine_protein_df, chemdiv_df, enamine_df, life_chemicals_df, robin_df]
sorted_datasets = sorted(datasets, key=lambda x: x.shape[0], reverse=True)  # Sort by size

# Check for duplicates again
combined_df = pd.concat(sorted_datasets)
duplicates_check = combined_df[combined_df.duplicated('SMILES', keep=False)]
if duplicates_check.empty:
    print("No duplicates found.")
else:
    print("Duplicates still exist.")


# which dataset has the most duplicates?
duplicates_check['source'].value_counts()


# check if True if not create dir named 'no_duplicates'
if not os.path.exists('../data/no_duplicates/v2'):
    os.makedirs('../data/no_duplicates/v2')

combined_df

## v4 are deleted duplicates fully from everything

# create function for ECFP calculation. Function will use rdkit's AllChem.GetMorganFingerprintAsBitVect() function and it will compute ECFP4 fingerprints with radius 2, ECFP6 fingerprints with radius 3
# ,ECFP8 fingerprints with radius 4 and ECFP10 fingerprints with radius 5. Function will create new columns for each ECFP and it will return dataframe with new columns.

def compute_ECFP(df, smiles_column, radius):
    df['ECFP'+str(radius*2)] = df[smiles_column].apply(lambda x: GetMorganFingerprintAsBitVect(MolFromSmiles(x),radius, nBits=2048))
    
    return df

compute_ECFP(combined_df, 'SMILES', 3)

# check for duplicates in SMILES column 
combined_df[combined_df.duplicated('SMILES')]

# split the dataset to 5 different datasets
chemdiv_df = combined_df[combined_df['source'] == 'chemdiv']
enamine_df = combined_df[combined_df['source'] == 'enamine']
enamine_protein_df = combined_df[combined_df['source'] == 'enamine_protein']
life_chemicals_df = combined_df[combined_df['source'] == 'life_chemicals']
robin_df = combined_df[combined_df['source'] == 'robin']


# reset indexes of each dataset
for dataset in [chemdiv_df, enamine_df, enamine_protein_df, life_chemicals_df, robin_df]:
    dataset.reset_index(drop=True, inplace=True)
    

# create a folder for ECFP datasets
if not os.path.exists('../data/ECFP_datasets'):
    os.makedirs('../data/ECFP_datasets')


# save ECFP datasets as pickle files
chemdiv_df.to_pickle('../../../data/ECFP_datasets/chemdiv_ECFP6_v4_dedup_all.pkl')
enamine_df.to_pickle('../../../data/ECFP_datasets/enamine_ECFP6_v4_dedup_all.pkl')
enamine_protein_df.to_pickle('../../../data/ECFP_datasets/enamine_protein_ECFP6_v4_dedup_all.pkl')
life_chemicals_df.to_pickle('../../../data/ECFP_datasets/life_chemicals_ECFP6_v4_dedup_all.pkl')
robin_df.to_pickle('../../../data/ECFP_datasets/robin_ECFP6_v4_dedup_all.pkl')
