==> notebooks/analysis/set1/04_v2_distance_matrix_compute_final.ipynb <==
import os
from rdkit import RDLogger
RDLogger.DisableLog('rdApp.error')
from rdkit import DataStructs
import numpy as np
import pandas as pd
from tqdm.auto import tqdm
tqdm.pandas()


# create directory for saving similarity matrices
if not os.path.exists('../output/distance_matrices/v2'):
    os.makedirs('../output/distance_matrices/v2')

    
n_protein = 8000
n_rna = 2000





enamine_protein_div6 = pd.read_pickle('../data/diversity_picking/ECFP6_v2/enamine_protein_ECFP6_MaxMin_v3.pkl').head(n_protein)
chemdiv_div6 = pd.read_pickle('../data/diversity_picking/ECFP6_v2/chemdiv_rna_ECFP6_MaxMin_v3.pkl').head(n_rna)
enamine_div6 = pd.read_pickle('../data/diversity_picking/ECFP6_v2/enamine_rna_ECFP6_MaxMin_v3.pkl').head(n_rna)
life_chemicals_div6 = pd.read_pickle('../data/diversity_picking/ECFP6_v2/life_chemicals_rna_ECFP6_MaxMin_v3.pkl').head(n_rna)
robin_df_div6 = pd.read_pickle('../data/diversity_picking/ECFP6_v2/robin_rna_ECFP6_MaxMin_v3.pkl').head(n_rna)



# Length of all dataframes
print(len(enamine_protein_div6), len(chemdiv_div6), len(enamine_div6), len(life_chemicals_div6), len(robin_df_div6))

# Combine datasets into one DataFrame
combined_df = pd.concat([enamine_protein_div6, chemdiv_div6, enamine_div6, life_chemicals_div6, robin_df_div6])

# Sort datasets by size (largest to smallest)
datasets = [enamine_protein_div6, chemdiv_div6, enamine_div6, life_chemicals_div6, robin_df_div6]
sorted_datasets = sorted(datasets, key=lambda x: x.shape[0], reverse=True)

# Finding all unique duplicates across datasets
all_duplicates = combined_df[combined_df.duplicated('smiles', keep=False)]

# Remove duplicates in order from largest to smallest, except the smallest
for i in range(len(sorted_datasets) - 1):  # Skip the smallest dataset
    dataset = sorted_datasets[i]
    duplicates_to_remove = all_duplicates[~all_duplicates['source'].isin([dataset['source'].iloc[0]])]['smiles']
    sorted_datasets[i] = dataset[~dataset['smiles'].isin(duplicates_to_remove)]

# Extracting the updated datasets
enamine_protein_div6, chemdiv_div6, enamine_div6, life_chemicals_div6, robin_df_div6 = sorted_datasets

# Now the datasets are deduplicated in the desired order

combined_df = pd.concat([
    enamine_protein_div6, chemdiv_div6, enamine_div6, 
    life_chemicals_div6, robin_df_div6
])

# Finding duplicates in the 'SMILES' column
duplicates = combined_df[combined_df.duplicated('smiles', keep=False)]

# Creating a summary of duplicates between each pair of datasets
summary = {}
for dataset1 in combined_df['source'].unique():
    for dataset2 in combined_df['source'].unique():
        if dataset1 != dataset2:
            pair = tuple(sorted([dataset1, dataset2]))
            if pair not in summary:
                duplicates_in_pair = duplicates[
                    (duplicates['source'] == dataset1) | 
                    (duplicates['source'] == dataset2)
                ]
                count = duplicates_in_pair['smiles'].nunique()
                summary[pair] = count

# Displaying the summary
for pair, count in summary.items():
    print(f"Duplicates between {pair[0]} and {pair[1]}: {count}")


# Combine the dataframes
combined_df6 = pd.concat([enamine_protein_div6, chemdiv_div6, enamine_div6, life_chemicals_div6, robin_df_div6], ignore_index=True)


n = combined_df6.shape[0]
ecfp_column = ['ECFP6']
df_list = []

for ecfp_col in tqdm(ecfp_column, desc="Calculating distance matrices"):
    if ecfp_col == 'ECFP6':
        combined_df = combined_df6

    distance_matrix = np.zeros((n, n))

    for i in range(n):
        if i % 100 == 0:
            print(i)
        for j in range(i, n):
            distance_matrix[i, j] = 1 - DataStructs.TanimotoSimilarity(combined_df[ecfp_col][i], combined_df[ecfp_col][j])
            distance_matrix[j, i] = distance_matrix[i, j]

    df = pd.DataFrame({'mol': combined_df['mol'], 'source': combined_df['source'], ecfp_col: distance_matrix.tolist()})
    df_list.append(df)

    file_name = f'../output/distance_matrices/v2/distance_matrix_{ecfp_col}.pkl'
    df.to_pickle(file_name)

# Concatenate the DataFrames for all ECFP columns
combined_df = pd.concat(df_list, axis=1)

# load pickle
dist_mat_ = pd.read_pickle('../output/distance_matrices/v2/distance_matrix_ECFP6.pkl')
dist_mat_
